split submodule for scicrunch
